Plutus.MerkleTree

Description

A purely functional implementation of MerkleTrees that is suitable for usage on-chain. Note however that the construction of MerkleTree and membership proofs are still expected to happen *off-chain* while only the proof verification should be done on-chain.

Note that this module is meant to used as a qualified import, for example:

import qualified Plutus.MerkleTree as MT

MerkleTree

data MerkleTree

A MerkleTree representation, suitable for on-chain manipulation. Construction of the merkle tree shouldn't be done by hand, but via fromList.

Constructors

MerkleEmpty

MerkleNode Hash MerkleTree MerkleTree

MerkleLeaf Hash BuiltinByteString

fromList :: [BuiltinByteString] -> MerkleTree

Construct a MerkleTree from a list of serialized data as BuiltinByteString.

Note that, while this operation is doable on-chain, it is expensive and preferably done off-chain.

toList :: MerkleTree -> [BuiltinByteString]

Deconstruct a MerkleTree back to a list of elements.

>>> toList (fromList xs) == xs
True

rootHash :: MerkleTree -> Hash

Obtain the root hash of a MerkleTree. In particular we have:

>>> (mt == mt') == (rootHash mt == rootHash mt')
True

null :: MerkleTree -> Bool

Return true if the MerkleTree is empty.

>>> null mt == (size mt == 0)
True

size :: MerkleTree -> Integer

Total numbers of leaves in the tree.

Proof

type Proof = [Either Hash Hash]

A membership Proof. The type is meant to be opaque.

mkProof :: BuiltinByteString -> MerkleTree -> Maybe Proof

Construct a membership Proof from an element and a MerkleTree. Returns Nothing if the element isn't a member of the tree to begin with.

member :: BuiltinByteString -> Hash -> Proof -> Bool

Check whether a element is part of a MerkleTree using only its root hash and a Proof. The proof is guaranteed to be in log(n) of the size of the tree, which is why we are interested in such data-structure in the first place.

Hash

newtype Hash

A type for representing hash digests.

Constructors

Hash BuiltinByteString

hash :: BuiltinByteString -> Hash

Computes a SHA-256 hash of a given BuiltinByteString message.

combineHash :: Hash -> Hash -> Hash

Combines two hashes digest into a new one. This is effectively a new hash digest of the same length.